+++ nacc_attribute_deriver/attributes/nacc/modules/uds/__init__.py
@@ -3,5 +3,6 @@ from .form_a1 import *
 from .form_b1 import *
 from .form_b6 import *
 from .form_b8 import *
+from .form_b9 import *
 from .form_d1 import *
 from .form_header import *