+++ nacc_attribute_deriver/attributes/nacc/modules/uds/form_b6.py
@@ -17,7 +17,6 @@ class UDSFormB6Attribute(AttributeCollection):
 
     def __init__(self, table: SymbolTable):
         self.__uds = UDSNamespace(table)
-        self.__formver = self.__uds.normalized_formver()
         self.__submitted = self.__uds.get_value("b6sub", int) == 1
 
     GDS_VARS: frozenset[str] = frozenset([