+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Documentation of release versions of `nacc-attribute-deriver`
 * Adds enforcement of expected types when grabbing variables from the raw data
 * Adds rest of NP variables
 * Adds rest of UDS A1, A2 derived variables
+* Adds UDS BX forms
+* Adds ListOperation (insert order is retained)
 
 
 ## 1.4.0